Move readers and loaders to confidence.io
